+++ internal/auth/code.go
@@ -33,7 +33,7 @@ const (
 
 type Code struct {
 	gorm.Model
-	Value   string    `gorm:"unique_index:idx_code_value"`
+	Value   string    `gorm:"uniqueIndex:idx_code_value"`
 	Expires time.Time `gorm:"index:idx_code_expires"`
 	Token   string
 }